+++ OLMoE.swift/Model/HuggingFaceError.swift
@@ -1,0 +1,13 @@
+//
+//  HuggingFaceError.swift
+//  OLMoE.swift
+//
+//  Created by Ken Adamson on 11/17/24.
+//
+
+
+public enum HuggingFaceError: Error {
+    case network(statusCode: Int)
+    case noFilteredURL
+    case urlIsNilForSomeReason
+}
