+++ OLMoE.swift/Model/Model.swift
@@ -1,0 +1,59 @@
+//
+//  Model.swift
+//  OLMoE.swift
+//
+//  Created by Ken Adamson on 11/17/24.
+//
+import llama
+
+extension Model {
+    public var endToken: Token { llama_token_eos(self) }
+    public var newLineToken: Token { llama_token_nl(self) }
+
+    public func shouldAddBOS() -> Bool {
+        let addBOS = llama_add_bos_token(self);
+        guard !addBOS else {
+            return llama_vocab_type(self) == LLAMA_VOCAB_TYPE_SPM
+        }
+        return addBOS
+    }
+
+    public func decodeOnly(_ token: Token) -> String {
+        var nothing: [CUnsignedChar] = []
+        return decode(token, with: &nothing)
+    }
+
+    public func decode(_ token: Token, with multibyteCharacter: inout [CUnsignedChar]) -> String {
+        var bufferLength = 16
+        var buffer: [CChar] = .init(repeating: 0, count: bufferLength)
+        let actualLength = Int(llama_token_to_piece(self, token, &buffer, Int32(bufferLength), 0, false))
+        guard 0 != actualLength else { return "" }
+        if actualLength < 0 {
+            bufferLength = -actualLength
+            buffer = .init(repeating: 0, count: bufferLength)
+            llama_token_to_piece(self, token, &buffer, Int32(bufferLength), 0, false)
+        } else {
+            buffer.removeLast(bufferLength - actualLength)
+        }
+        if multibyteCharacter.isEmpty, let decoded = String(cString: buffer + [0], encoding: .utf8) {
+            return decoded
+        }
+        multibyteCharacter.append(contentsOf: buffer.map { CUnsignedChar(bitPattern: $0) })
+        guard let decoded = String(data: .init(multibyteCharacter), encoding: .utf8) else { return "" }
+        multibyteCharacter.removeAll(keepingCapacity: true)
+        return decoded
+    }
+
+    public func encode(_ text: borrowing String) -> [Token] {
+        let addBOS = true
+        let count = Int32(text.cString(using: .utf8)!.count)
+        var tokenCount = count + 1
+        let cTokens = UnsafeMutablePointer<llama_token>.allocate(capacity: Int(tokenCount)); defer { cTokens.deallocate() }
+        tokenCount = llama_tokenize(self, text, count, cTokens, tokenCount, addBOS, false)
+        let tokens = (0..<Int(tokenCount)).map { cTokens[$0] }
+
+        print("Encoded tokens: \(tokens)")  // Add this line to log the resulting tokens
+
+        return tokens
+    }
+}
