+++ OLMoE.swift/Model/Role.swift
@@ -1,0 +1,12 @@
+//
+//  Role.swift
+//  OLMoE.swift
+//
+//  Created by Ken Adamson on 11/17/24.
+//
+
+
+public enum Role {
+    case user
+    case bot
+}
