+++ OLMoE.swift/Views/SpinnerView.swift
@@ -11,12 +11,12 @@ struct SpinnerView: View {
     let color: Color
     let size: CGFloat
     @State private var isAnimating = false
-    
+
     init(color: Color = .accentColor, size: CGFloat = 24) {
         self.color = color
         self.size = size
     }
-    
+
     var body: some View {
         Circle()
             .trim(from: 0, to: 0.7)
