+++ OLMoE.swift/extensions/String.swift
@@ -1,0 +1,28 @@
+//
+//  String.swift
+//  OLMoE.swift
+//
+//  Created by Ken Adamson on 11/17/24.
+//
+
+import Foundation
+
+extension String {
+    func matches(in content: String) throws -> [String] {
+        let pattern = try NSRegularExpression(pattern: self)
+        let range = NSRange(location: 0, length: content.utf16.count)
+        let matches = pattern.matches(in: content, range: range)
+        return matches.map { match in String(content[Range(match.range, in: content)!]) }
+    }
+    func hasMatch(in content: String) throws -> Bool {
+        let pattern = try NSRegularExpression(pattern: self)
+        let range = NSRange(location: 0, length: content.utf16.count)
+        return pattern.firstMatch(in: content, range: range) != nil
+    }
+    func firstMatch(in content: String) throws -> String? {
+        let pattern = try NSRegularExpression(pattern: self)
+        let range = NSRange(location: 0, length: content.utf16.count)
+        guard let match = pattern.firstMatch(in: content, range: range) else { return nil }
+        return String(content[Range(match.range, in: content)!])
+    }
+}
