+++ OLMoE.swift/extensions/URL.swift
@@ -1,0 +1,57 @@
+//
+//  URL.swift
+//  OLMoE.swift
+//
+//  Created by Ken Adamson on 11/17/24.
+//
+import Foundation
+
+extension URL {
+    @backDeployed(before: iOS 16)
+    public func appending(path: String) -> URL {
+        appendingPathComponent(path)
+    }
+    @backDeployed(before: iOS 16)
+    public static var modelsDirectory: URL {
+        let paths = FileManager.default.urls(for: .applicationSupportDirectory, in: .userDomainMask)
+        let url = paths[0].appendingPathComponent("Models")
+        try? FileManager.default.createDirectory(at: url, withIntermediateDirectories: true)
+
+        // Exclude from backup
+        do {
+            var mutableURL = url
+            var resourceValues = URLResourceValues()
+            resourceValues.isExcludedFromBackup = true
+            try mutableURL.setResourceValues(resourceValues)
+        } catch {
+            print("Error excluding from backup: \(error)")
+        }
+
+        return url
+    }
+    public var exists: Bool { FileManager.default.fileExists(atPath: path) }
+    public func getData() async throws -> Data {
+        let (data, response) = try await URLSession.shared.data(from: self)
+        let statusCode = (response as! HTTPURLResponse).statusCode
+        guard statusCode / 100 == 2 else { throw HuggingFaceError.network(statusCode: statusCode) }
+        return data
+    }
+    public func downloadData(to destination: URL, _ updateProgress: @escaping (Double) -> Void) async throws {
+        var observation: NSKeyValueObservation!
+        let url: URL = try await withCheckedThrowingContinuation { continuation in
+            let task = URLSession.shared.downloadTask(with: self) { url, response, error in
+                if let error { return continuation.resume(throwing: error) }
+                guard let url else { return continuation.resume(throwing: HuggingFaceError.urlIsNilForSomeReason) }
+                let statusCode = (response as! HTTPURLResponse).statusCode
+                guard statusCode / 100 == 2 else { return continuation.resume(throwing: HuggingFaceError.network(statusCode: statusCode)) }
+                continuation.resume(returning: url)
+            }
+            observation = task.progress.observe(\.fractionCompleted) { progress, _ in
+                updateProgress(progress.fractionCompleted)
+            }
+            task.resume()
+        }
+        _ = observation
+        try FileManager.default.moveItem(at: url, to: destination)
+    }
+}
